requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.15.2
pytz==2022.2.1
sqlparse==0.5.0
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
